+++ src/kernel/framework/tests/mod.rs
@@ -33,6 +33,7 @@ pub mod test_hvfs;
 pub mod test_hvfs_ext;
 pub mod test_ipc;
 pub mod test_mm;
+pub mod test_uds;
 pub mod test_new_features;
 pub mod test_proc;
 pub mod test_pwm;
@@ -345,6 +346,7 @@ pub fn test_runner_init() {
     test_mm::register_mm_tests();
     test_vfs::register_vfs_tests();
     test_ipc::register_ipc_tests();
+    test_uds::register_uds_tests();
     test_devfs::register_devfs_tests();
     test_proc::register_proc_tests();
     test_new_features::register_new_tests();